change a lot

+++ engine_java/src/main/java/com/alibabacloud/polar_race/engine/common/Location.java
@@ -2,11 +2,11 @@ package com.alibabacloud.polar_race.engine.common;
 
 public class Location {
 
-    public long offset;
+    public int offset;
 
-    public long fileNo;
+    public int fileNo;
 
-    public Location(long offset, long fileNo) {
+    public Location(int offset, int fileNo) {
         this.offset = offset;
         this.fileNo = fileNo;
     }
